Removed fb4_ module, as it is mostly useless now.
