+++ bench/bench_fbx.c
@@ -25,7 +25,6 @@
  *
  * Benchmarks for extensions defined over binary fields.
  *
- * @version $Id$
  * @ingroup bench
  */
 
@@ -152,155 +151,6 @@ static void arith2(void) {
 	fb2_free(c);
 }
 
-static void memory4(void) {
-	fb4_t a[BENCH];
-
-	BENCH_SMALL("fb4_null", fb4_null(a[i]));
-
-	BENCH_SMALL("fb4_new", fb4_new(a[i]));
-	for (int i = 0; i < BENCH; i++) {
-		fb4_free(a[i]);
-	}
-
-	for (int i = 0; i < BENCH; i++) {
-		fb4_new(a[i]);
-	}
-	BENCH_SMALL("fb4_free", fb4_free(a[i]));
-
-	(void)a;
-}
-
-static void util4(void) {
-	fb4_t a, b;
-
-	fb4_null(a);
-	fb4_null(b);
-
-	fb4_new(a);
-	fb4_new(b);
-
-	BENCH_BEGIN("fb4_copy") {
-		fb4_rand(a);
-		BENCH_ADD(fb4_copy(b, a));
-	}
-	BENCH_END;
-
-	BENCH_BEGIN("fb4_neg") {
-		fb4_rand(a);
-		BENCH_ADD(fb4_neg(b, a));
-	}
-	BENCH_END;
-
-	BENCH_BEGIN("fb4_zero") {
-		fb4_rand(a);
-		BENCH_ADD(fb4_zero(a));
-	}
-	BENCH_END;
-
-	BENCH_BEGIN("fb4_is_zero") {
-		fb4_rand(a);
-		BENCH_ADD((void)fb4_is_zero(a));
-	}
-	BENCH_END;
-
-	BENCH_BEGIN("fb4_rand") {
-		BENCH_ADD(fb4_rand(a));
-	}
-	BENCH_END;
-
-	BENCH_BEGIN("fb4_cmp") {
-		fb4_rand(a);
-		fb4_rand(b);
-		BENCH_ADD(fb4_cmp(b, a));
-	}
-	BENCH_END;
-
-	fb4_free(a);
-	fb4_free(b);
-}
-
-static void arith4(void) {
-	fb4_t a, b, c;
-	bn_t d;
-
-	fb4_new(a);
-	fb4_new(b);
-	fb4_new(c);
-	bn_new(d);
-
-	BENCH_BEGIN("fb4_add") {
-		fb4_rand(a);
-		fb4_rand(b);
-		BENCH_ADD(fb4_add(c, a, b));
-	}
-	BENCH_END;
-
-	BENCH_BEGIN("fb4_sub") {
-		fb4_rand(a);
-		fb4_rand(b);
-		BENCH_ADD(fb4_sub(c, a, b));
-	}
-	BENCH_END;
-
-	BENCH_BEGIN("fb4_mul") {
-		fb4_rand(a);
-		fb4_rand(b);
-		BENCH_ADD(fb4_mul(c, a, b));
-	}
-	BENCH_END;
-
-	BENCH_BEGIN("fb4_mul_dxd") {
-		fb4_rand(a);
-		fb4_rand(b);
-		BENCH_ADD(fb4_mul_dxd(c, a, b));
-	}
-	BENCH_END;
-
-	BENCH_BEGIN("fb4_mul_dxs") {
-		fb4_rand(a);
-		fb4_rand(b);
-		BENCH_ADD(fb4_mul_dxs(c, a, b));
-	}
-	BENCH_END;
-
-	BENCH_BEGIN("fb4_mul_sxs") {
-		fb4_rand(a);
-		fb4_rand(b);
-		BENCH_ADD(fb4_mul_sxs(c, a, b));
-	}
-	BENCH_END;
-
-	BENCH_BEGIN("fb4_sqr") {
-		fb4_rand(a);
-		BENCH_ADD(fb4_sqr(c, a));
-	}
-	BENCH_END;
-
-	BENCH_BEGIN("fb4_inv") {
-		fb4_rand(a);
-		BENCH_ADD(fb4_inv(c, a));
-	}
-	BENCH_END;
-
-	BENCH_BEGIN("fb4_frb") {
-		fb4_rand(a);
-		BENCH_ADD(fb4_frb(c, a));
-	}
-	BENCH_END;
-
-	BENCH_BEGIN("fb4_exp") {
-		fb4_rand(a);
-		bn_rand(d, BN_POS, FB_BITS);
-		BENCH_ADD(fb4_exp(c, a, d));
-	}
-	BENCH_END;
-
-	fb4_free(a);
-	fb4_free(b);
-	fb4_free(c);
-	bn_free(d);
-}
-
 int main(void) {
 	if (core_init() != STS_OK) {
 		core_clean();
@@ -321,14 +171,6 @@ int main(void) {
 
 	util_banner("Arithmetic:", 1);
 	arith2();
-
-	util_banner("Quartic extension:", 0);
-	util_banner("Utilities", 1)
-	memory4();
-	util4();
-
-	util_banner("Arithmetic:", 1);
-	arith4();
 
 	core_clean();
 	return 0;
