--- src/fbx/relic_fbx_exp.c
@@ -1,63 +1,0 @@
-/*
- * RELIC is an Efficient LIbrary for Cryptography
- * Copyright (C) 2007-2015 RELIC Authors
- *
- * This file is part of RELIC. RELIC is legal property of its developers,
- * whose names are not listed here. Please refer to the COPYRIGHT file
- * for contact information.
- *
- * RELIC is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- *
- * RELIC is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with RELIC. If not, see <http://www.gnu.org/licenses/>.
- */
-
-/**
- * @file
- *
- * Implementation of exponentiation in binary field extensions.
- *
- * @version $Id$
- * @ingroup fbx
- */
-
-#include "relic_core.h"
-#include "relic_fbx.h"
-
-/*============================================================================*/
-/* Public definitions                                                         */
-/*============================================================================*/
-
-void fb4_exp(fb4_t c, fb4_t a, bn_t b) {
-	fb4_t t;
-
-	fb4_null(t);
-
-	TRY {
-		fb4_new(t);
-
-		fb4_copy(t, a);
-
-		for (int i = bn_bits(b) - 2; i >= 0; i--) {
-			fb4_sqr(t, t);
-			if (bn_get_bit(b, i)) {
-				fb4_mul(t, t, a);
-			}
-		}
-		fb4_copy(c, t);
-	}
-	CATCH_ANY {
-		THROW(ERR_CAUGHT);
-	}
-	FINALLY {
-		fb4_free(t);
-	}
-}
